requirements.txt
asgiref==3.8.1
Django==5.1.3
django-debug-toolbar==4.4.6
pillow==11.0.0
pip==24.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
setuptools==75.5.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.0
